third_party/WebKit/LayoutTests/fast/dom/Range/range-intersectsNode-exception.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<!-- hidden area to create the ranges being tested -->
<div  style="visibility: hidden">
    <div id=a1>a1
        <div id=b1>b1</div>
    </div>
</div>

<script>
description("Test for Bug: 75799 - Calling intersectsNode on a detached range should throw");
var range = document.createRange();
range.selectNode(document.getElementById("a1"));
range.detach();
shouldThrow("range.intersectsNode(document.getElementById('b1'))", '"InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable."');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
